refactor(workers): moved all worker related code to a "workers" folder
